Draft implementation for core functionality for resource limited jobs (#38)
Many small changes were required, but the majority were mechanical
changes. With the exception of the `createJob`/`scheduleJob` family of
function, this should be mostly backwards compatible with existing
user code that does not use "internal" functions. Further discussion
about how to handle backwards compatibility is needed.

Note also that this commit by itself is NOT compatible with existing
databases. A future commit will address migration from existing to new
database definitions.

The intention is for this feature to have a negligable impact on
performance when the resource limiting features are not being used,
and only a minor impact when they are used. However, performance
testing has not yet been done.

Material changes inclue:

- New resource table
- New `resource_id` column in to job table
- New `cfgDefaultResourceLimit` configuration value
- New `createResourceJob` and `scheduleResourceJob` functions
- New `TableNames` type to carry the now multiple values needed
- Updated existing functions and queries to account for these changes
- Updated dequeueing queries to implement resource concurrency limits
- New `DevelDb` module to the `devel` executable to help ad hoc testing

TODO:
- Resource record management functions
- Migration functionality (need to discuss database versioning)
- Unit tests
- Performance tests
- Documentation updates
- Resource management UI