Introduced `dropJobTables` and refactored `Migrations.hs`
Because there are unit tests that delete tabes, and now that there are
multiple table names in play, it seemed useful to explicitly implement
a deletion function. Becuase this deserved a comment, I also added
comments to the other externally useful functions in this module, and
rearranged the file a bit to highlight these defintions.