Push failed jobs to end of the queue
Order jobpollingsql by attempts first so that failed jobs go to end
of the queue.

The reason is that failed jobs are likely to fail again.
In our case we have some really heavy jobs that consume a lot
of resources, interspersed with lighter ones.
If one of the heavy job fails, it's better some other job
gets a chance to run first. Odd jobs can work on that heavy one again
when it's less busy (during night for example).