Add ability to kill ongoing job
This adds the ability to kill a running job from the web UI, by adding
an action to the menu for running jobs. When the "kill" button is
clicked, the job is marked for cancellation and an asynchronous poller
is responsible for reaping it later.

This adds a new status "Cancelled", which represents a job that has been
manually killed and is either waiting to be reaped, or has been reaped
and can be queued again.

To allow for cancelling job threads by ID, the internal runner now
tracks job IDs alongside the spawned threads.

Co-authored-by: Isaac van Bakel <isaac@supercede.com>