Fix resource checking when no jobs are locked
The way the resource usage calculating function was defined had a slight
bug, where `NULL` would be returned if no jobs were currently running
which used the resource. This propagated up to the function which
checked if resources were free, and meant that the job system would
always queue up the first job to use a resource, since it couldn't
comapre `NULL` to the resource's limit, and defaulted to `true` when
deciding whether to run it.

This adds some tests for cases where there should be no jobs running and
the resource limits should prevent the runner from starting a job. These
cases come from real-world examples which were failing due to this bug.