Add configurable retry backoff
The default retry backoff is not very long, and assumes that the retry
limit is quite high. This is less useful for configurations where the
retry limit is low and immediate retries can hurt more than they help.

This adds the retry backoff to the backend config, so that the developer
can set their own backoff strategy. The config defaults to the current
approach, so there should be no change in behaviour for existing users.