Implement onJobTimeout callback
This fixes the behaviour of job timeouts to allow the `onJobTimeout`
callback to be defined. Previously, the timeout thread would itself
cancel the job thread before throwing a timeout exception - but due to
the use of `waitEitherCancel`, the cancellation exception would be
rethrown to the call to `runJob`, rather than the `TimeoutException`.
This meant the handler for `TimeoutException` was never being called.

With these changes, the `TimeoutException` is itself the one thrown from
`waitEitherCancel`, and that function is instead responsible for
cancelling the job thread. This allows the `retryOrFail` code to act
differently for `TimeoutException`s, calling `onJobTimeout` rather than
`onJobFailed`.

This also fixes some warnings in the tests.